refactor(db): replace old DataContext implementation with updated configuration for SQL Server
